requirements.txt
python-dotenv
langchain
chromadb
tiktoken
langchain-google-genai
